classmanager/pyproject.toml
[tool.poetry]
name = "application"
version = "0.1.0"
description = ""
authors = ["Tobiloba Ogundiyan <114788338+6ogunt48@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
asyncpg = "^0.28.0"
tortoise-orm = "^0.20.0"
gunicorn = "^21.2.0"
aerich = "^0.7.2"
pydantic = {extras = ["email"], version = "^2.4.0"}
passlib = "^1.7.4"
bcrypt = "^4.0.1"




[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
flake8 = "^6.1.0"
safety = "^2.3.5"
isort = "^5.12.0"
black = "^23.9.1"
pytest-cov = "^4.1.0"
pytest = "^7.4.2"
httpx = "^0.25.0"


[tool.aerich]
tortoise_orm = "application.db.database_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
